Create eggs.html
(I realized the previous commit only affected the recipes folder, this
commit will affect the entire project as I originally intended)
Created eggs.html and added an image of scrambled eggs, a description
of the scrambled eggs recipe, the required ingredients and the required
steps to complete the recipe. Added a link on eggs.html to index.html.
Also added a link on index.html to eggs.html.